ScrollLock: Locking fix + resolve console warning
This update fixes the ScrollLock mechanism to work with the newer browser +
React handling for `wheel` events.

The event listener is now added directly to the DOM node via `addEventListener`,
rather than tapping into the `onWheel` callback from React, which is now
considered "passive", and no longer provides us with the correct `eventTarget`.

`@helpscout/cyan` was added to help make it (much MUCH) easier to write
`wheel` event based tests, since these are now fired outside of React.
