+++ config/jest/setupTests.js
@@ -1,5 +1,6 @@
 import Enzyme from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
+import { setupTests } from '@helpscout/cyan'
 
 Enzyme.configure({ adapter: new Adapter() })
 
@@ -10,3 +11,5 @@ beforeEach(() => {
 afterEach(() => {
   window.requestAnimationFrame.mockRestore()
 })
+
+setupTests()
