+++ package.json
@@ -130,6 +130,7 @@
     "@babel/runtime": "^7.0.0",
     "@helpscout/artboard": "0.1.12",
     "@helpscout/colorway": "0.0.11",
+    "@helpscout/cyan": "0.0.14",
     "@helpscout/helix": "^0.1.0",
     "@helpscout/prestart": "^0.0.9",
     "@helpscout/stats": "0.0.5",
@@ -183,7 +184,7 @@
     "jest-cli": "^23.6.0",
     "jest-watch-typeahead": "^0.2.0",
     "mkdirp": "0.5.1",
-    "node-sass": "^4.9.2",
+    "node-sass": "4.11.0",
     "np": "4.0.2",
     "nyc": "12.0.2",
     "object-assign": "4.1.1",
