+++ src/components/ScrollLock/ScrollLock.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react'
+import * as ReactDOM from 'react-dom'
 import { noop } from '../../utilities/other'
 import propConnect from '../PropProvider/propConnect'
 import { COMPONENT_KEY, handleWheelEvent } from './ScrollLock.utils'
 
-type Props = {
+export interface Props {
   children?: any
   direction: 'x' | 'y'
   isDisabled: boolean
@@ -18,30 +19,41 @@ export class ScrollLock extends React.PureComponent<Props> {
     stopPropagation: false,
     onWheel: noop,
   }
+  node: Element | null
+
+  componentDidMount() {
+    if (this.canRender()) {
+      this.node = ReactDOM.findDOMNode(this) as Element
+      /* istanbul ignore else */
+      if (this.node) {
+        this.node.addEventListener('wheel', this.handleWheelEvent)
+      }
+    }
+  }
+
+  componentWillUnmount() {
+    if (this.node) {
+      this.node.removeEventListener('wheel', this.handleWheelEvent)
+    }
+  }
+
+  canRender() {
+    return !!this.props.children
+  }
+
+  handleWheelEvent = (event: any) => {
+    const { direction, isDisabled, onWheel, stopPropagation } = this.props
+    onWheel(event)
+
+    if (isDisabled) return
+
+    handleWheelEvent(event, direction, stopPropagation)
+  }
 
   render() {
-    const {
-      children,
-      direction,
-      isDisabled,
-      onWheel,
-      stopPropagation,
-    } = this.props
+    if (!this.canRender()) return null
 
-    if (!children) {
-      return null
-    }
-
-    const child = React.Children.only(children)
-    const events = {
-      onWheel: (event: any) => {
-        handleWheelEvent(event, direction, stopPropagation)
-        onWheel(event)
-        if (child.props.onWheel) child.props.onWheel(event)
-      },
-    }
-
-    return isDisabled ? child : React.cloneElement(child, events)
+    return React.Children.only(this.props.children)
   }
 }
